Implement learning rate decrease and save model.
The network now has a slow decrease in the learning rate to hopefully speed up
the rate of optimisation and reach peak accuracy faster.

The best epoch and accuracy are also saved to allow for picking up of the model
at a later stage to decide whether to reuse this model for further testing or
improve it later with better parameters.

Still todo:
- Try include Network in Network layer, this will involve writing a network layer
have a look at this paper: https://arxiv.org/abs/1312.4400 for further details.
- Reduce number of layered samples, this may involve creating a labeled dataset
loader and an unlabeled equivalent.
- Possible on going consideration is increasing dataset to include time series,
further data points - will this require a time series gan such as those used
for music synthesis?